atualização

+++ project/financeiro/models.py
@@ -6,40 +6,78 @@ from django.db.models.signals import post_save, post_delete
 from django.dispatch import receiver
 import datetime
 
+
 from compra.models import Compra
 from despesa.models import Despesa
 from receita.models import Receita
 from venda.models import FormaPagamento, Venda
 
 today = datetime.date.today
+time = datetime.datetime.now()
+date = time.strftime("%d-%m-%y")
+hour = time.strftime("%H")
+minute = time.strftime("%M")
+seconds = time.strftime("%S")
+printar = f"{hour}H{minute}m{seconds}s"
 
 
-def path_and_rename(instance, filename):
-    upload_to = "financeiro"
-    time = datetime.datetime.now()
-    hour = time.strftime("%H")
-    minute = time.strftime("%M")
-    printar = f"{hour}H{minute}m"
-    if instance.venda.id:
-        filename = "{}_{}.jpg".format(instance.venda.vendedor, printar)
+def path_and_receita_boleto(instance, filename):
+    ext = filename.split(".")
+    if instance.venda:
+        filename = "{}_{}_{}.{}".format(instance.venda.vendedor, date, printar, ext[-1])
     else:
-        filename = "{}_{}.jpg".format(instance.receita.vendedor, printar)
-    return os.path.join(upload_to, filename)
+        filename = "{}_{}_{}.{}".format(instance.receita.nome, date, printar, ext[-1])
+    return os.path.join(filename)
+
+
+def path_and_receita_comprovante(instance, filename):
+    ext = filename.split(".")
+    if instance.venda:
+        filename = "{}_{}_{}.{}".format(instance.venda.vendedor, date, printar, ext[-1])
+    else:
+        filename = "{}_{}_{}.{}".format(instance.receita.nome, date, printar, ext[-1])
+    return os.path.join(filename)
+
+
+def path_and_despesa_boleto(instance, filename):
+    ext = filename.split(".")
+    if instance.compra:
+        filename = "{}_{}_{}.{}".format(
+            instance.compra.fornecedor, date, printar, ext[-1]
+        )
+    else:
+        filename = "{}_{}_{}.{}".format(instance.despesa.nome, date, printar, ext[-1])
+    return os.path.join(filename)
+
+
+def path_and_despesa_comprovante(instance, filename):
+    ext = filename.split(".")
+    if instance.compra:
+        filename = "{}_{}_{}.{}".format(
+            instance.compra.fornecedor, date, printar, ext[-1]
+        )
+    else:
+        filename = "{}_{}_{}.{}".format(instance.despesa.nome, date, printar, ext[-1])
+    return os.path.join(filename)
 
 
 class ContaReceber(models.Model):
-    venda = models.ForeignKey(Venda, on_delete=models.RESTRICT, null=True)
+    venda = models.ForeignKey(Venda, on_delete=models.CASCADE, null=True)
     parcela = models.CharField(
         max_length=11, null=True, blank=True
     )  # número da parcela
-    receita = models.ForeignKey(Receita, on_delete=models.RESTRICT, null=True)
+    receita = models.ForeignKey(Receita, on_delete=models.CASCADE, null=True)
     valor = models.DecimalField(max_digits=10, decimal_places=2, default=0.0)
     datadocumento = models.DateField(default=today)
     datavencimento = models.DateField("Data do vencimento")
     datapagamento = models.DateField(blank=True, null=True)
-    imagem = models.ImageField(upload_to=path_and_rename, null=True, blank=True)
+    boleto = models.FileField(upload_to=path_and_receita_boleto, null=True, blank=True)
+    comprovante = models.FileField(
+        upload_to=path_and_receita_comprovante, null=True, blank=True
+    )
     responsavel = models.CharField(max_length=255, null=True, blank=True)
     formapgto = models.ForeignKey(FormaPagamento, on_delete=models.RESTRICT)
+    dados_boleto = models.JSONField(default=dict, null=True, blank=True)
     detalhes = models.CharField(max_length=300, null=True, blank=True)
 
     class Meta:
@@ -61,13 +99,14 @@ def comissaoAdd(sender, instance, **kwargs):
     if instance.venda:
         percent_comis = Decimal(0.03)
         comissao_vlr = instance.valor * percent_comis
-    if not comissao:
-        Comissao.objects.create(parcela=instance, comissao=comissao_vlr)
+        if not comissao:
+            Comissao.objects.create(parcela=instance, comissao=comissao_vlr)
 
 
 @receiver(post_delete, sender=ContaReceber)
 def comissaoDel(sender, instance, **kwargs):
-    Comissao.objects.get(parcela=instance).delete()
+    if Comissao.objects.filter(parcela=instance):
+        Comissao.objects.get(parcela=instance).delete()
 
 
 class Comissao(models.Model):
@@ -80,27 +119,36 @@ class Comissao(models.Model):
         verbose_name_plural = "Comissões"
 
     def __str__(self):
-        return f"{self.parcela.venda.cliente} - {self.parcela.venda} - parcela: {self.parcela}"
+        return str(self.parcela)
 
     def comissao_edit(self):
         return _("financeiro:updateDataComissao", args=[self.pk])
 
 
 class ContaPagar(models.Model):
-    compra = models.ForeignKey(Compra, on_delete=models.CASCADE, null=True)
-    despesa = models.ForeignKey(Despesa, on_delete=models.CASCADE, null=True)
-    valor = models.DecimalField(max_digits=10, decimal_places=2, default=0.0)
+    compra = models.ForeignKey(Compra, on_delete=models.CASCADE, null=True, blank=True)
+    despesa = models.ForeignKey(
+        Despesa, on_delete=models.CASCADE, null=True, default=""
+    )
+    valor = models.DecimalField(max_digits=10, decimal_places=2, default=0.00)
     datadocumento = models.DateField(default=today)
     datavencimento = models.DateField()
     datapagamento = models.DateField(null=True, blank=True)
+    boleto = models.FileField(upload_to=path_and_despesa_boleto, null=True, blank=True)
+    comprovante = models.FileField(
+        upload_to=path_and_despesa_comprovante, null=True, blank=True
+    )
     formapgto = models.ForeignKey(FormaPagamento, on_delete=models.RESTRICT)
     detalhes = models.CharField(max_length=300, null=True, blank=True)
 
     def __str__(self):
         if self.despesa:
-            return "%s - %s" % (self.despesa.nome, self.valor)
+            return "%s - %s" % (self.despesa.nome, self.datadocumento)
         if self.compra:
-            return "%s - %s" % (self.compra.fornecedor, self.valor)
+            return "%s - %s" % (self.compra.fornecedor, self.datadocumento)
+
+    def parcela_gau_delete(self):
+        return _("financeiro:parcelaDelete", args=[self.pk])
 
 
 @receiver(post_save, sender=ContaPagar)
